PreFinalVersion/requirements.txt
requests
pandas
openpyxl
openai
python-dotenv
tensorflow
tensorflow_hub
tensorflow_text